django_backend/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.6
channels==4.0
djangorestframework-simplejwt==5.2.2
celery==5.2.7
redis==4.5.3
django-cors-headers==4.2.2
gunicorn==20.1.0
